backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-whois==0.9.5
requests==2.32.3
pydantic-settings==2.7.1
python-multipart==0.0.20

# --- ML stack ---
# torch + transformers power SMS classification (DistilBERT) and the
# ResNet-18 backbone for image screening. Pin loosely — the Hugging Face
# ecosystem moves fast and narrow pins cause weekly breakage.
torch>=2.4,<3.0
torchvision>=0.19,<1.0
transformers>=4.45,<5.0
datasets>=3.0
accelerate>=1.0

# Explainability
lime>=0.2.0.1
scikit-learn>=1.5

# Image processing
Pillow>=10.0
numpy>=1.26

# Evaluation / thesis artifacts — confusion matrices, reliability
# diagrams, ablation tables. Used by scripts/evaluate_*.py and
# scripts/ablation_*.py; not needed for runtime inference.
matplotlib>=3.8

# Tests
pytest>=8.0
httpx>=0.27
